src/Makefile: separate DESTDIR and PREFIX
They are different things. DESTDIR is used by distributions to install
to a temporary path that will be turned into a package. PREFIX is used
to decide if one is going to install to / (root), /usr or another
desired path.

Automated build systems from distributions like Void Linux automatically
set PREFIX and DESTDIR to the proper values.